refactor: CalorieWorker 관련 클래스 패키지 변경 및 의존성 주입 구조 개선
