+++ android/app/src/main/java/com/mulkkam/domain/repository/HealthRepository.kt
@@ -12,7 +12,6 @@ interface HealthRepository {
     suspend fun hasPermissions(permissions: Set<String>): Boolean
 
     companion object {
-        const val DEFAULT_CHECK_CALORIE_INTERVAL_HOURS: Long = 2L
         const val SECONDS_IN_TWO_HOURS: Long = 3600 * 2L
     }
 }
