+++ android/app/src/main/java/com/mulkkam/ui/main/MainViewModel.kt
@@ -4,6 +4,8 @@ import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.mulkkam.di.RepositoryInjection.healthRepository
+import com.mulkkam.di.WorkInjection.calorieScheduler
+import com.mulkkam.domain.work.CalorieScheduler.Companion.DEFAULT_CHECK_CALORIE_INTERVAL_HOURS
 import kotlinx.coroutines.launch
 
 class MainViewModel : ViewModel() {
@@ -19,5 +21,9 @@ class MainViewModel : ViewModel() {
 
     fun updateHealthPermissionStatus(isGranted: Boolean) {
         _isHealthPermissionGranted.value = isGranted
+    }
+
+    fun scheduleCalorieCheck() {
+        calorieScheduler.scheduleCalorieCheck(DEFAULT_CHECK_CALORIE_INTERVAL_HOURS)
     }
 }
