+++ android/app/src/main/java/com/mulkkam/util/work/CalorieSchedulerImpl.kt
@@ -1,15 +1,15 @@
-package com.mulkkam.util
+package com.mulkkam.util.work
 
 import androidx.work.ExistingPeriodicWorkPolicy
 import androidx.work.PeriodicWorkRequestBuilder
 import androidx.work.WorkManager
-import com.mulkkam.di.WorkInjection.workManager
+import com.mulkkam.domain.work.CalorieScheduler
 import java.util.concurrent.TimeUnit
 
 class CalorieSchedulerImpl(
     private val workManager: WorkManager,
-) {
-    fun scheduleCalorieCheck(intervalHours: Long) {
+) : CalorieScheduler {
+    override fun scheduleCalorieCheck(intervalHours: Long) {
         val request = PeriodicWorkRequestBuilder<CalorieWorker>(intervalHours, TimeUnit.HOURS).build()
         workManager.enqueueUniquePeriodicWork(
             CHECK_CALORIE_WORK_NAME,
@@ -20,15 +20,5 @@ class CalorieSchedulerImpl(
 
     companion object {
         private const val CHECK_CALORIE_WORK_NAME = "CHECK_CALORIE_WORK"
-
-        @Volatile
-        private var instance: CalorieSchedulerImpl? = null
-
-        fun getOrCreate(): CalorieSchedulerImpl =
-            instance ?: synchronized(this) {
-                instance ?: CalorieSchedulerImpl(workManager).also {
-                    instance = it
-                }
-            }
     }
 }
