+++ android/app/src/main/java/com/mulkkam/util/work/CalorieWorker.kt
@@ -1,4 +1,4 @@
-package com.mulkkam.util
+package com.mulkkam.util.work
 
 import android.content.Context
 import android.util.Log
