+++ android/app/src/main/java/com/mulkkam/util/work/WorkerFactory.kt
@@ -1,4 +1,4 @@
-package com.mulkkam.util
+package com.mulkkam.util.work
 
 import android.content.Context
 import androidx.work.ListenableWorker
@@ -6,7 +6,7 @@ import androidx.work.WorkerFactory
 import androidx.work.WorkerParameters
 import com.mulkkam.domain.repository.HealthRepository
 
-class CalorieWorkerFactory(
+class WorkerFactory(
     private val healthRepository: HealthRepository,
 ) : WorkerFactory() {
     override fun createWorker(
